lambda_functions/requirements.txt
fastapi==0.110.0
mangum==0.17.0
openai==1.10.0
python-dotenv==1.0.5
boto3==1.28.38
PyPDF2==3.0.1
langtrace-python-sdk==0.0.32
crewai==0.28.5
pydantic==2.6.0 